More cl testing and fixed glaring issue with cl filtering

+++ server/src/api/v1/handlers/sp.rs
@@ -207,7 +207,7 @@ pub async fn check_for_valid_score(
     };
 
     values.score_delta = Some(cl[0].score - score);
-    values.previous_id = Some(cl[0].id as i32); // TODO: Fix this
+    values.previous_id = Some(cl[0].id);
     // Assuming there is a PB History, there must be other scores, this should return a valid list of ranked maps.
     let cl_ranked = SpMap::get_sp_map_page(pool, map_id).await.unwrap();
     for (i, entry) in cl_ranked.iter().enumerate() {
